requirement.txt
# Python version requirement: 3.11.11
# Install with: pyenv install 3.11.11 && pyenv local 3.11.11

# Core dependencies
python-dotenv==1.0.0
qdrant-client==1.12.0

# Embedding and ML
FlagEmbedding>=1.2.0
transformers>=4.21.0
torch>=1.9.0

# LangChain ecosystem
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langgraph>=0.2.0
langchain-openai>=0.1.0

# File processing
pandas>=2.1.0
numpy>=1.24.0

# Utilities
tqdm>=4.66.0

# MCP (Model Context Protocol) dependencies
mcp>=1.0.0
fastmcp>=0.1.0
langchain-mcp-adapters>=0.1.0